requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
defusedxml==0.7.1
etelemetry==0.3.1
filelock==3.20.2
firebase_admin==7.1.0
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.60.1
fpdf2==2.8.5
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.2
magic-filter==1.0.12
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pyxnat==1.6.4
rdflib==7.5.0
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
Werkzeug==3.1.4
yarl==1.22.0
